added more generic modifier hints
