+++ Source/controls/modifier_hints.cpp
@@ -18,67 +18,6 @@ extern std::optional<CelSprite> pSBkIconCels;
 
 namespace {
 
-int SpaceWidth()
-{
-	static const int Result = GetLineWidth(" ");
-	return Result;
-}
-
-/** The number of spaces between left and right hints. */
-constexpr int MidSpaces = 5;
-
-/** Vertical distance between text lines. */
-constexpr int LineHeight = 25;
-
-/** Horizontal margin of the hints circle from panel edge. */
-constexpr int CircleMarginX = 16;
-
-/** Distance between the panel top and the circle top. */
-constexpr int CircleTop = 101;
-
-/** Spell icon side size. */
-constexpr int IconSize = 37;
-
-/** Spell icon text right margin. */
-constexpr int IconSizeTextMarginRight = 3;
-
-/** Spell icon text top margin. */
-constexpr int IconSizeTextMarginTop = 2;
-
-struct CircleMenuHint {
-	CircleMenuHint(bool isDpad, const char *top, const char *right, const char *bottom, const char *left)
-	    : isDpad(isDpad)
-	    , top(top)
-	    , topW(GetLineWidth(top))
-	    , right(right)
-	    , rightW(GetLineWidth(right))
-	    , bottom(bottom)
-	    , bottomW(GetLineWidth(bottom))
-	    , left(left)
-	    , leftW(GetLineWidth(left))
-	    , xMid(leftW + SpaceWidth() * MidSpaces / 2)
-	{
-	}
-
-	[[nodiscard]] int Width() const
-	{
-		return 2 * xMid;
-	}
-
-	bool isDpad;
-
-	const char *top;
-	int topW;
-	const char *right;
-	int rightW;
-	const char *bottom;
-	int bottomW;
-	const char *left;
-	int leftW;
-
-	int xMid;
-};
-
 bool IsTopActive(const CircleMenuHint &hint)
 {
 	if (hint.isDpad)
@@ -185,15 +124,17 @@ void DrawStartModifierMenu(const Surface &out)
 	if (!start_modifier_active)
 		return;
 
-#ifdef __UWP__
-	static const CircleMenuHint DPad(/*isDpad=*/true, /*top=*/_("Quests"), /*right=*/_("Inv"), /*bottom=*/_("Spells"), /*left=*/_("Char"));
-	DrawCircleMenuHint(out, DPad, { PANEL_LEFT + CircleMarginX, PANEL_TOP - CircleTop });
-#else
-	static const CircleMenuHint DPad(/*isDpad=*/true, /*top=*/_("Menu"), /*right=*/_("Inv"), /*bottom=*/_("Map"), /*left=*/_("Char"));
-	static const CircleMenuHint Buttons(/*isDpad=*/false, /*top=*/"", /*right=*/"", /*bottom=*/_("Spells"), /*left=*/_("Quests"));
-	DrawCircleMenuHint(out, DPad, { PANEL_LEFT + CircleMarginX, PANEL_TOP - CircleTop });
-	DrawCircleMenuHint(out, Buttons, { PANEL_LEFT + PANEL_WIDTH - Buttons.Width() - CircleMarginX, PANEL_TOP - CircleTop });
-#endif
+	CircleMenuHint *hint = nullptr;
+
+	if (GetStartModifierLeftCircleMenuHint(&hint)) {
+		DrawCircleMenuHint(out, *hint, { PANEL_LEFT + CircleMarginX, PANEL_TOP - CircleTop });
+		delete hint;
+	}
+
+	if (GetStartModifierRightCircleMenuHint(&hint)) {
+		DrawCircleMenuHint(out, *hint, { PANEL_LEFT + PANEL_WIDTH - hint->Width() - CircleMarginX, PANEL_TOP - CircleTop });
+		delete hint;
+	}
 }
 
 void DrawSelectModifierMenu(const Surface &out)
@@ -201,14 +142,17 @@ void DrawSelectModifierMenu(const Surface &out)
 	if (!select_modifier_active)
 		return;
 
-#ifndef __UWP__
-	if (sgOptions.Controller.bDpadHotkeys) {
-		static const CircleMenuHint DPad(/*isDpad=*/true, /*top=*/"F6", /*right=*/"F8", /*bottom=*/"F7", /*left=*/"F5");
-		DrawSpellsCircleMenuHint(out, DPad, { PANEL_LEFT + CircleMarginX, PANEL_TOP - CircleTop });
+	CircleMenuHint *hint = nullptr;
+
+	if (GetSelectModifierLeftCircleMenuHint(&hint)) {
+		DrawSpellsCircleMenuHint(out, *hint, { PANEL_LEFT + CircleMarginX, PANEL_TOP - CircleTop });
+		delete hint;
 	}
-#endif
-	static const CircleMenuHint Spells(/*isDpad=*/false, "F6", "F8", "F7", "F5");
-	DrawSpellsCircleMenuHint(out, Spells, { PANEL_LEFT + PANEL_WIDTH - IconSize * 3 - CircleMarginX, PANEL_TOP - CircleTop });
+
+	if (GetSelectModifierRightCircleMenuHint(&hint)) {
+		DrawSpellsCircleMenuHint(out, *hint, { PANEL_LEFT + PANEL_WIDTH - IconSize * 3 - CircleMarginX, PANEL_TOP - CircleTop });
+		delete hint;
+	}
 }
 
 } // namespace
